requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
pypdf==5.1.0
python-docx==1.1.2
reportlab==4.2.5
requests==2.32.3
python-dotenv==1.0.1
chardet==5.2.0
